% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensor_data.R
\name{get_sensor_data}
\alias{get_sensor_data}
\title{Retrieves current sensor information for a single sensor (defined by 'sensor_index').}
\usage{
get_sensor_data(
  sensor_index,
  fields = NULL,
  api_key = NULL,
  geometry = FALSE,
  ...
)
}
\arguments{
\item{sensor_index}{A single numeric value corresponding to a specific PurpleAir sensor}

\item{fields}{Optional; a character vector of fields specified in the \href{https://api.purpleair.com/#api-sensors-get-sensors-data}{PurpleAir API documentation}. If left blank, will default to all fields specified in the documentation}

\item{api_key}{If not set with purpleair_api_key()}

\item{geometry}{Optional (default FALSE); if TRUE, an sf object will be returned.}

\item{...}{}
}
\value{
A POINT sf object
}
\description{
Retrieves current sensor information for a single sensor (defined by 'sensor_index').
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
get_sensors_data(nc, c("name", "pm2.5", "humidity"))
}
