% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensors_data.R
\name{get_sensors_data}
\alias{get_sensors_data}
\title{Retrieves a list with the latest data of multiple sensors matching the provided parameters.}
\usage{
get_sensors_data(aoi = NULL, fields, api_key = NULL, ...)
}
\arguments{
\item{aoi}{Optional; an sf object defining an area of interest}

\item{fields}{A character vector of fields specified in the \href{https://api.purpleair.com/#api-sensors-get-sensors-data}{PurpleAir API documentation}}

\item{api_key}{If not previously set with purpleair_api_key()}

\item{...}{}
}
\value{
A POINT sf object
}
\description{
Retrieves a list with the latest data of multiple sensors matching the provided parameters.
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
get_sensors_data(nc, c("name", "pm2.5", "humidity"))
}
